+++ nifi-commons/nifi-record/src/test/java/org/apache/nifi/serialization/record/TestDataTypeUtils.java
@@ -498,11 +498,11 @@ public class TestDataTypeUtils {
         assertInstanceOf(Byte[].class, bytes);
         assertNotNull(bytes);
         Byte[] b = (Byte[]) bytes;
-        assertEquals((long) 72, (long) b[0], "Conversion from String to byte[] failed");  // H
-        assertEquals((long) 101, (long) b[1], "Conversion from String to byte[] failed"); // e
-        assertEquals((long) 108, (long) b[2], "Conversion from String to byte[] failed"); // l
-        assertEquals((long) 108, (long) b[3], "Conversion from String to byte[] failed"); // l
-        assertEquals((long) 111, (long) b[4], "Conversion from String to byte[] failed"); // o
+        assertEquals(72, (long) b[0], "Conversion from String to byte[] failed");  // H
+        assertEquals(101, (long) b[1], "Conversion from String to byte[] failed"); // e
+        assertEquals(108, (long) b[2], "Conversion from String to byte[] failed"); // l
+        assertEquals(108, (long) b[3], "Conversion from String to byte[] failed"); // l
+        assertEquals(111, (long) b[4], "Conversion from String to byte[] failed"); // o
     }
 
     @Test
